data-pipeline/requirements.txt
numpy==1.24.3
pandas==2.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
azure-storage-blob==12.18.3
opencv-python==4.8.0.76
scikit-learn==1.3.0
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.1.0
mypy==1.5.1
requests==2.31.0
python-jose==3.3.0 